backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
fastapi==0.110.0
h11==0.14.0
idna==3.6
loadenv==0.1.1
motor==3.3.2
outcome==1.3.0.post0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.2
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.36.3
termcolor==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
wsproto==1.2.0
xlwt==1.3.0
